+++ cysyphus_frontend/src/COLORS.css
@@ -1,0 +1,5 @@
+:root {
+    --maincolor: #333333;
+    --mainhover: #4a4a4a;
+    --accentcolor: #f5d7a1;
+}
