+++ cysyphus_frontend/src/widgets_styling/NavBar.css
@@ -1,0 +1,25 @@
+@import "../COLORS.css";
+
+.NavBarStyling {
+    background-color: var(--maincolor);
+    width: 100%;
+}
+
+.LinkStyle, .LoginButton {
+    color: var(--accentcolor);
+    cursor: pointer;
+
+    font-family: Arial;
+    text-align: center;
+    text-decoration: none;
+    font-size: 18px;
+
+    padding: 15px 32px;
+    display: inline-block;
+}
+.LoginButton {
+    float: right;
+}
+.LinkStyle:hover, .LoginButton:hover {
+    background-color: var(--mainhover);
+}
